Add WasDeal property to OrderItem for deal tracking
Introduced a new boolean property `WasDeal` in the `OrderItem` entity to indicate if an item was part of a deal, with a default value of `false`. Updated the `OrderItem` class and removed previous default values for `Quantity` and `PricePerUnit`. Modified the `OrderService` to set `WasDeal` based on product features and added logging for featured deal purchases.